+++ contracts/plugins/assets/morpho-aave/MorphoSelfReferentialCollateral.sol
@@ -10,8 +10,6 @@ import { shiftl_toFix, FIX_ONE, FLOOR, FixLib, CEIL } from "../../../libraries/F
 
 // solhint-enable max-line-length
 
-// solhint-enable max-line-length
-
 /**
  * @title MorphoSelfReferentialCollateral
  * @notice Collateral plugin for a Morpho pool with self referential collateral, like WETH
@@ -41,6 +39,7 @@ contract MorphoSelfReferentialCollateral is AppreciatingFiatCollateral {
     }
 
     /// Can revert, used by other contract functions in order to catch errors
+    /// Should NOT be manipulable by MEV
     /// @return low {UoA/tok} The low price estimate
     /// @return high {UoA/tok} The high price estimate
     /// @return pegPrice {target/ref}