+++ test/registries/DAOFeeRegistry.test.ts
@@ -9,6 +9,7 @@ import { whileImpersonating } from '../utils/impersonation'
 import {
   DAOFeeRegistry,
   ERC20Mock,
+  RoleRegistry,
   TestIDistributor,
   TestIRevenueTrader,
   TestIMain,
@@ -28,6 +29,7 @@ describeP1('DAO Fee Registry', () => {
   let rsrTrader: TestIRevenueTrader
 
   let feeRegistry: DAOFeeRegistry
+  let roleRegistry: RoleRegistry
 
   beforeEach(async () => {
     ;[owner, other] = await ethers.getSigners()
@@ -35,15 +37,19 @@ describeP1('DAO Fee Registry', () => {
     // Deploy fixture
     ;({ distributor, main, rToken, rsr, rsrTrader } = await loadFixture(defaultFixture))
 
+    const RoleRegistryFactory = await ethers.getContractFactory('RoleRegistry')
+    roleRegistry = await RoleRegistryFactory.connect(owner).deploy()
+
     const DAOFeeRegistryFactory = await ethers.getContractFactory('DAOFeeRegistry')
-    feeRegistry = await DAOFeeRegistryFactory.connect(owner).deploy(await owner.getAddress())
+    feeRegistry = await DAOFeeRegistryFactory.connect(owner).deploy(
+      roleRegistry.address,
+      await owner.getAddress()
+    )
+
     await main.connect(owner).setDAOFeeRegistry(feeRegistry.address)
   })
 
   describe('Deployment', () => {
-    it('should set the owner correctly', async () => {
-      expect(await feeRegistry.owner()).to.eq(await owner.getAddress())
-    })
     it('fee should begin zero and assigned to owner', async () => {
       const feeDetails = await feeRegistry.getFeeDetails(rToken.address)
       expect(feeDetails.recipient).to.equal(owner.address)
@@ -52,37 +58,20 @@ describeP1('DAO Fee Registry', () => {
     })
   })
 
-  describe('Ownership', () => {
-    it('Should be able to change owner', async () => {
-      expect(await feeRegistry.owner()).to.eq(await owner.getAddress())
-      await feeRegistry.connect(owner).transferOwnership(other.address)
-      expect(await feeRegistry.owner()).to.eq(await other.getAddress())
-      await expect(feeRegistry.connect(owner).setFeeRecipient(owner.address)).to.be.revertedWith(
-        'Ownable: caller is not the owner'
-      )
-      await expect(feeRegistry.connect(owner).setDefaultFeeNumerator(bn('100'))).to.be.revertedWith(
-        'Ownable: caller is not the owner'
-      )
-      await expect(
-        feeRegistry.connect(owner).setRTokenFeeNumerator(rToken.address, bn('100'))
-      ).to.be.revertedWith('Ownable: caller is not the owner')
-    })
-  })
-
   describe('Negative cases', () => {
     it('Should not allow calling setters by anyone other than owner', async () => {
-      await expect(feeRegistry.connect(other).setFeeRecipient(owner.address)).to.be.revertedWith(
-        'Ownable: caller is not the owner'
-      )
-      await expect(feeRegistry.connect(other).setDefaultFeeNumerator(bn('100'))).to.be.revertedWith(
-        'Ownable: caller is not the owner'
-      )
+      await expect(
+        feeRegistry.connect(other).setFeeRecipient(owner.address)
+      ).to.be.revertedWithCustomError(feeRegistry, 'DAOFeeRegistry__InvalidCaller')
+      await expect(
+        feeRegistry.connect(other).setDefaultFeeNumerator(bn('100'))
+      ).to.be.revertedWithCustomError(feeRegistry, 'DAOFeeRegistry__InvalidCaller')
       await expect(
         feeRegistry.connect(other).setRTokenFeeNumerator(rToken.address, bn('100'))
-      ).to.be.revertedWith('Ownable: caller is not the owner')
-      await expect(feeRegistry.connect(other).resetRTokenFee(rToken.address)).to.be.revertedWith(
-        'Ownable: caller is not the owner'
-      )
+      ).to.be.revertedWithCustomError(feeRegistry, 'DAOFeeRegistry__InvalidCaller')
+      await expect(
+        feeRegistry.connect(other).resetRTokenFee(rToken.address)
+      ).to.be.revertedWithCustomError(feeRegistry, 'DAOFeeRegistry__InvalidCaller')
     })
 
     it('Should not allow setting fee recipient to zero address', async () => {
@@ -205,9 +194,9 @@ describeP1('DAO Fee Registry', () => {
               expect(await rsr.balanceOf(owner.address)).to.equal(0)
               await distributor.connect(signer).distribute(rsr.address, amt)
 
-              // Expected returned amount is for the fee times 5/3 to account for rev share split
-              const expectedAmt = amt.mul(defaultFee).div(bn('1e4')).mul(5).div(3)
-              expect(await rsr.balanceOf(owner.address)).to.equal(expectedAmt)
+              const feeShares = bn('1e4').mul(defaultFee).div(bn('1e4').sub(defaultFee))
+              const expectedAmt = amt.mul(feeShares).div(feeShares.add(6000))
+              expect(await rsr.balanceOf(owner.address)).to.be.closeTo(expectedAmt, 10000)
             })
           })
         }